+++ NEWS.md
@@ -1,5 +1,9 @@
 # Changes in version 1.1.2
 
+* adding volume instead of area for MERFISH.
+* refactoring log2CountArea to log2DensitySignal
+
+
 # Changes in version 1.1.1
 
 * fixing minor bugs in AspectRatio internal computation for technology missing it.